Delete actions.js
to replace it with slice
